added pdf end to end preprocessing test
