+++ tests/example_data/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+
+EXAMPLE_DATA_DIR = os.path.dirname(__file__)
